Move mypage component (#178)
