+++ frontend/app/routes/home.me._index/components/MyLoansListComponent.tsx
@@ -1,10 +1,10 @@
 import { Stack } from '@mantine/core';
 import { SerializeFrom } from '@remix-run/cloudflare';
 import { getLoansResponse } from 'client/client';
+import ErrorComponent from '~/components/common/error/ErrorComponent';
+import ContentsHeader from '~/components/common/pagination/ContentsHeader';
+import PaginationComponent from '~/components/common/pagination/PaginationComponent';
 import { PaginationProps } from '~/types/pagination';
-import ErrorComponent from '../common/error/ErrorComponent';
-import ContentsHeader from '../common/pagination/ContentsHeader';
-import PaginationComponent from '../common/pagination/PaginationComponent';
 import LoanCards from './LoanCards';
 
 interface MyLoansListComponentProps {
