+++ frontend/app/routes/home.me._index/components/MyPageComponent.tsx
@@ -3,8 +3,8 @@ import { SerializeFrom } from '@remix-run/cloudflare';
 import { getLoansResponse } from 'client/client';
 import { User } from 'client/client.schemas';
 import { FaUser } from 'react-icons/fa';
+import BreadCrumbsComponent from '~/components/common/breadcrumbs/BreadCrumbsComponent';
 import { PaginationProps } from '~/types/pagination';
-import BreadCrumbsComponent from '../common/breadcrumbs/BreadCrumbsComponent';
 import MyLoansListComponent from './MyLoansListComponent';
 import MyProfileComponent from './MyProfileComponent';
